Application/recomendacao/requirements.txt
Flask==3.0.3
pandas==2.2.3
python-dotenv==1.0.1
redis==5.1.1
